backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
scikit-learn
numpy
pandas
python-dotenv
